made separate reducer for the selectedDate
