+++ src/redux/date/date.actions.js
@@ -1,0 +1,5 @@
+import { createAction } from "@reduxjs/toolkit";
+
+const getSelectedDate = createAction("date/getSelectedDate");
+
+export default getSelectedDate;
