+++ src/redux/date/date.selectors.js
@@ -1,0 +1,3 @@
+const getSelectedDate = (state) => state.selectedDate.selectedDate;
+
+export { getSelectedDate };
