+++ src/redux/transactions/transactions.reducer.js
@@ -69,18 +69,18 @@ const brief = createReducer(
   }
 );
 
-const selectedDateReducer = createReducer(
-  { selectedDate: "" },
-  {
-    [actions.getSelectedDate]: (state, { payload }) => {
-      return { selectedDate: payload };
-    },
-  }
-);
+// const selectedDateReducer = createReducer(
+//   { selectedDate: "" },
+//   {
+//     [actions.getSelectedDate]: (state, { payload }) => {
+//       return { selectedDate: payload };
+//     },
+//   }
+// );
 
 export default combineReducers({
   // balanceReducer,
   list: transactionsReducer,
   brief,
-  selectedDate: selectedDateReducer,
+  // selectedDate: selectedDateReducer,
 });
